+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## [1.3.0] - 2026-02-24
+
+### Added
+
+- Added `lazy` behavior to `EnLogger`
+- Added `dispose` and `close` methods to `EnLogger`
+
+### Changed
+
+- chore: improved documentation (README and code documentation)
+
 ## [1.2.1] - 2026-01-17
 
 ### Changed